+++ src/main/java/at/ac/fhcampuswien/battleship/Controller.java
@@ -1,0 +1,5 @@
+package at.ac.fhcampuswien.battleship;
+
+public class Controller
+{
+}
