+++ src/main/java/at/ac/fhcampuswien/battleship/ImageShip.java
@@ -1,4 +1,4 @@
-package sample;
+package at.ac.fhcampuswien.battleship;
 
 import javafx.event.EventHandler;
 import javafx.scene.image.Image;
