+++ src/main/java/at/ac/fhcampuswien/battleship/ShipPart.java
@@ -1,4 +1,4 @@
-package sample;
+package at.ac.fhcampuswien.battleship;
 
 public class ShipPart {
