+++ src/main/java/module-info.java
@@ -1,5 +1,5 @@
 module battleship {
     requires transitive javafx.controls;
     requires javafx.media;
-    exports sample;
+    exports at.ac.fhcampuswien.battleship;
 }
