+++ src/main/java/package-info.java
@@ -1,0 +1,1 @@
+package at.ac.fhcampuswien.battleship;
